Rewrite a bit of ReloadedGraphic.
The class that launched a thousand rewrites.

The problem with the last iteration was that the class's width and
height didn't respect what they should have been with the mask. So it
would report crazy large width and heights, rather than the size of the
smaller mask. And there was no way to get around it because you can't
override width/height.